Track the comments preceding a setting in the ConfigOrigin for the value.
The rule is that any block of comments uninterrupted by another token
or blank line, goes with the following array element or object field.

The comments are shown in the output of ConfigValue.render().

They are also available in the ConfigOrigin API for custom use.

Comments that don't precede an array element of object field
get discarded, they are not available anywhere.