remove NeedsFullResolver internal checked exception
I think this was a relic of some earlier idea about how to do
things. It was used in one place that could just use the
public NotResolved exception, since it was converted to that
later anyhow.

To avoid error-message regression, we still catch NotResolved
and change it into a NotResolved with a different message in
certain cases.

Also clean up the wording of NotResolved in various places.
Probably NotResolved could just have a zero-parameters constructor.