Make ConfigException serializable
Done with a custom serialization because we don't want ConfigOrigin
to be serializable; if ConfigOrigin were serializable we'd have to
support its serialization remaining compatible across releases.
And there's no known use for ConfigOrigin being serializable.

Not intending to support ConfigException serialization back compat
across releases, because that probably isn't useful or expected.