Cleanup of ConfigDocumentParser and tests
Clean up ConfigDocumentParser tests by using the `intercept`
method to handle exception testing, moving some tests, and
removing a test that was re-testing already tested functionality.

Fix bug wherein parsing of a single value would throw an exception
when the value is a map with a key that has an object value and no
separator if the default ConfParserOptions where used. Add a test
to validate this fix.