Fix the Map spec implementation for ConfigObject.
According to the Map specification, methods keySet, values and entrySet are backed by the map, so changes to the map are reflected in the collection, and vice-versa. ConfigObject interface extends Map, so it must conform the specification. Also, ConfigObject declared as immutable, so keySet, values and entrySet must return immutable collections.