Merge adjacent objects in delayed merge stacks
The reason this is needed is that it prevents trying to
resolve those objects prior to merging them.

Previously, if we had something like this:

 foo : ""
 foo : { bar : ${undefined} }
 foo : { bar : 42 }

Then it was possible to end up with the value of foo as a delayed merge with
multiple adjacent objects in its stack.

We then would resolve each value in the stack, and only after that
merge all the values. But we need to merge objects BEFORE we try to
resolve them, in case some of the unresolved substitutions just
go away.

This was only happening with the bottom of the stack as a non-object,
since if the bottom was an object, we were merging the objects
as we went along anyway.