Support naive masking of config values
You can supply a regex pattern to match against Config key names that it
will mask it corresponding config value.  If the config value is a
config reference, it will keep the config reference.

This is a naive approach, and secrets will not be masked if the config
reference key is also not matching the maskRegex.

This is useful for printing configuration "safely" (read: best effort)
without exposing secrets.

First version for feature:
https://github.com/lightbend/config/issues/145